upd: remove #include main.h from headers
