+++ divi/src/masternode-payments.h
@@ -6,8 +6,10 @@
 #ifndef MASTERNODE_PAYMENTS_H
 #define MASTERNODE_PAYMENTS_H
 
+#ifndef BLOCK_REWARDS_H
+#include <BlockRewards.h>
+#endif // BLOCK_REWARDS_H
 #include "key.h"
-#include "main.h"
 #include "masternode.h"
 #include <boost/lexical_cast.hpp>
 #include "LotteryCoinstakes.h"
