+++ divi/src/masternodeman.h
@@ -8,7 +8,6 @@
 
 #include "base58.h"
 #include "key.h"
-#include "main.h"
 #include "masternode.h"
 #include "net.h"
 #include "sync.h"
