+++ divi/src/rpcwallet.cpp
@@ -25,6 +25,9 @@
 #include "json/json_spirit_value.h"
 #include "spork.h"
 #include <boost/assign/list_of.hpp>
+#ifndef BITCOIN_MAIN_H
+#include <main.h>
+#endif // BITCOIN_MAIN_H
 
 using namespace std;
 using namespace boost;
