+++ divi/src/walletdb.cpp
@@ -14,7 +14,9 @@
 #include "util.h"
 #include "utiltime.h"
 #include "wallet.h"
+#ifndef BITCOIN_MAIN_H
 #include <main.h>
+#endif
 #include <boost/filesystem.hpp>
 #include <boost/foreach.hpp>
 #include <boost/scoped_ptr.hpp>
